Fix inheritance complication
