+++ src/Endpoints/CollectionEndpointAbstract.php
@@ -3,6 +3,7 @@
 namespace Mollie\Api\Endpoints;
 
 use Mollie\Api\Resources\BaseCollection;
+use Mollie\Api\Resources\ResourceFactory;
 
 abstract class CollectionEndpointAbstract extends EndpointAbstract
 {
@@ -15,4 +16,32 @@ abstract class CollectionEndpointAbstract extends EndpointAbstract
      * @return BaseCollection
      */
     abstract protected function getResourceCollectionObject($count, $_links);
+
+    /**
+     * Get a collection of objects from the REST API.
+     *
+     * @param string $from The first resource ID you want to include in your list.
+     * @param int $limit
+     * @param array $filters
+     *
+     * @return BaseCollection
+     * @throws \Mollie\Api\Exceptions\ApiException
+     */
+    protected function rest_list($from = null, $limit = null, array $filters)
+    {
+        $filters = array_merge(["from" => $from, "limit" => $limit], $filters);
+
+        $apiPath = $this->getResourcePath() . $this->buildQueryString($filters);
+
+        $result = $this->client->performHttpCall(self::REST_LIST, $apiPath);
+
+        /** @var BaseCollection $collection */
+        $collection = $this->getResourceCollectionObject($result->count, $result->_links);
+
+        foreach ($result->_embedded->{$collection->getCollectionResourceName()} as $dataResult) {
+            $collection[] = ResourceFactory::createFromApiResult($dataResult, $this->getResourceObject());
+        }
+
+        return $collection;
+    }
 }
