+++ src/Endpoints/EndpointAbstract.php
@@ -133,50 +133,11 @@ abstract class EndpointAbstract
     }
 
     /**
-     * Get a collection of objects from the REST API.
-     *
-     * @param string $from The first resource ID you want to include in your list.
-     * @param int $limit
-     * @param array $filters
-     *
-     * @return BaseCollection
-     * @throws ApiException
-     */
-    protected function rest_list($from = null, $limit = null, array $filters)
-    {
-        $filters = array_merge(["from" => $from, "limit" => $limit], $filters);
-
-        $apiPath = $this->getResourcePath() . $this->buildQueryString($filters);
-
-        $result = $this->client->performHttpCall(self::REST_LIST, $apiPath);
-
-        /** @var BaseCollection $collection */
-        $collection = $this->getResourceCollectionObject($result->count, $result->_links);
-
-        foreach ($result->_embedded->{$collection->getCollectionResourceName()} as $dataResult) {
-            $collection[] = ResourceFactory::createFromApiResult($dataResult, $this->getResourceObject());
-        }
-
-        return $collection;
-    }
-
-    /**
      * Get the object that is used by this API endpoint. Every API endpoint uses one type of object.
      *
      * @return BaseResource
      */
     abstract protected function getResourceObject();
-
-    /**
-     * Get the collection object that is used by this API endpoint. Every API
-     * endpoint uses one type of collection object.
-     *
-     * @param int $count
-     * @param \stdClass $_links
-     *
-     * @return CursorCollection
-     */
-    abstract protected function getResourceCollectionObject($count, $_links);
 
     /**
      * @param string $resourcePath
